Add bearing data and use better location provider

+++ core/src/main/java/com/wearfleet/core/events/BearingEvent.java
@@ -1,13 +1,13 @@
 package com.wearfleet.core.events;
 
 public class BearingEvent {
-    private final int bearing;
+    private final float bearing;
 
-    public BearingEvent(int bearing){
+    public BearingEvent(float bearing){
         this.bearing = bearing;
     }
 
-    public int getBearing() {
+    public float getBearing() {
         return bearing;
     }
 }
